2025.12.30

+++ User/foc/foc.h
@@ -24,7 +24,7 @@ void foc_set_target(foc_t *handle, float target_Id, float target_Iq, float targe
 /* 电机零点对齐 */
 void foc_alignment(foc_t *handle);
 
-void foc_speed_closed_loop(foc_t *handle, float angle_el, abc_t *i_abc, uint16_t speed_rpm);
+void foc_speed_closed_loop(foc_t *handle, float angle_el, abc_t *i_abc, float speed_rpm);
 
 void foc_open_loop(dq_t u_dq, float angle_rad_el);
